pyproject.toml
[project]
name = "migration-mongodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "pymongo",
    "python-dotenv",
]
